Make float <-> int casts explicit outside of test, qt, CFeeRate
Summary:
Backport of core PR11303
Only the commit 53a6590 is applicable for us.
https://github.com/bitcoin/bitcoin/pull/11303/commits/53a6590f496b25174c740927243bf8307541b0b9

Test Plan:
  make check

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3313

+++ src/rpc/net.cpp
@@ -183,7 +183,8 @@ static UniValue getpeerinfo(const Config &config,
         if (stats.dPingTime > 0.0) {
             obj.pushKV("pingtime", stats.dPingTime);
         }
-        if (stats.dMinPing < std::numeric_limits<int64_t>::max() / 1e6) {
+        if (stats.dMinPing <
+            static_cast<double>(std::numeric_limits<int64_t>::max()) / 1e6) {
             obj.pushKV("minping", stats.dMinPing);
         }
         if (stats.dPingWait > 0.0) {
